Fix on Moodles_TR.txt
Important fix on Wet moodle, which causes confusion and misleading information, especially for beginners. At the current version, first level of wet moodle suggests that only the rain is definitive cause. And the higher levels suggest a non-existent speed loss. 
Other than this spesific moodle, one minor misspell. 
And fixes to the appearance and categorization of the code to help with the any forecoming edits.